+++ src/actions/dynamoConfig.ts
@@ -3,196 +3,3 @@ import { generateClient } from "aws-amplify/data";
 
 export const DynamoClient = generateClient<Schema>();
 
-export const DemoData = [
-  {
-    userId: "user-001",
-    profileOwner: "user-001",
-    firstName: "Emma",
-    firstNameLower: "emma",
-    lastName: "Thompson",
-    lastNameLower: "thompson",
-    gender: "Female",
-    languages: ["English", "Spanish"],
-    phone: "+1-555-123-4567",
-    email: "emma.thompson@example.com",
-    preferredContact: "Email",
-    profilePhoto: "/images/home/meet-providers/person-placeholder-6.jpg",
-    address: "123 Maple Street",
-    city: "Toronto",
-    province: "Ontario",
-    postalCode: "M5V 2T6",
-    emergencyContactName: "James Thompson",
-    emergencyContactPhone: "+1-555-987-6543",
-    emergencyContactRelationship: "Brother",
-    profileTitle: "Registered Nurse",
-    bio: "Compassionate nurse with over 10 years of experience in pediatric care.",
-    yearsExperience: "5+ years",
-    yearsExperienceFloat: 5.5,
-    askingRate: 45.0,
-    rateType: "Hourly",
-    responseTime: "Within 24 hours",
-    servicesOffered: ["Dementia Care", "Personal Care"],
-    isProfileComplete: true,
-    lastUpdated: 1708650175742,
-    isPubliclyVisible: true,
-    availability: [
-      "2025-07-21:18",
-      "2025-07-21:19",
-      "2025-07-23:20",
-      "2025-07-26:17",
-      "2025-07-26:18"
-    ]
-  }, {
-    userId: "user-002",
-    profileOwner: "user-002",
-    firstName: "Liam",
-    firstNameLower: "liam",
-    lastName: "Patel",
-    lastNameLower: "patel",
-    gender: "Male",
-    languages: ["English", "Hindi", "Gujarati"],
-    phone: "+1-555-234-5678",
-    email: "liam.patel@example.com",
-    preferredContact: "Phone",
-    profilePhoto: "/images/home/meet-providers/person-placeholder-7.jpg",
-    address: "456 Oak Avenue",
-    city: "Vancouver",
-    province: "British Columbia",
-    postalCode: "V6B 3K9",
-    emergencyContactName: "Anita Patel",
-    emergencyContactPhone: "+1-555-876-5432",
-    emergencyContactRelationship: "Wife",
-    profileTitle: "Physiotherapist",
-    bio: "Experienced physiotherapist specializing in sports injury recovery.",
-    yearsExperience: "Less than 1 year",
-    yearsExperienceFloat: 0.5,
-    askingRate: 60.0,
-    rateType: "Hourly",
-    responseTime: "Within 12 hours",
-    servicesOffered: ["Personal Care", "Health Monitoring"],
-    isProfileComplete: true,
-    lastUpdated: 1708650175742,
-    isPubliclyVisible: true,
-    availability: [
-      "2025-07-24:08",
-      "2025-07-24:15",
-      "2025-07-27:12",
-      "2025-07-27:16",
-      "2025-07-29:10",
-      "2025-07-29:14"
-    ]
-  }, {
-    userId: "user-003",
-    profileOwner: "user-003",
-    firstName: "Sophie",
-    firstNameLower: "sophie",
-    lastName: "Chen",
-    lastNameLower: "chen",
-    gender: "Female",
-    languages: ["English", "Mandarin"],
-    phone: "+1-555-345-6789",
-    email: "sophie.chen@example.com",
-    preferredContact: "Email",
-    profilePhoto: "/images/home/meet-providers/person-placeholder-8.jpg",
-    address: "789 Pine Road",
-    city: "Calgary",
-    province: "Alberta",
-    postalCode: "T2P 4R5",
-    emergencyContactName: "Michael Chen",
-    emergencyContactPhone: "+1-555-765-4321",
-    emergencyContactRelationship: "Father",
-    profileTitle: "Occupational Therapist",
-    bio: "Dedicated occupational therapist helping clients regain independence.",
-    yearsExperience: "3-5 years",
-    yearsExperienceFloat: 3.5,
-    askingRate: 50.0,
-    rateType: "Hourly",
-    responseTime: "Within 24 hours",
-    servicesOffered: ["Health Monitoring", "Housekeeping"],
-    isProfileComplete: true,
-    lastUpdated: 1708650175742,
-    isPubliclyVisible: true,
-    availability: [
-      "2025-07-20:09",
-      "2025-07-20:10",
-      "2025-07-22:13",
-      "2025-07-22:14",
-      "2025-07-25:11"
-    ]
-  }, {
-    userId: "user-004",
-    profileOwner: "user-004",
-    firstName: "Noah",
-    firstNameLower: "noah",
-    lastName: "Rodriguez",
-    lastNameLower: "rodriguez",
-    gender: "Male",
-    languages: ["English", "Spanish", "French"],
-    phone: "+1-555-456-7890",
-    email: "noah.rodriguez@example.com",
-    preferredContact: "Phone",
-    profilePhoto: "/images/home/meet-providers/person-placeholder-9.jpg",
-    address: "101 Birch Lane",
-    city: "Montreal",
-    province: "Quebec",
-    postalCode: "H3B 2Y5",
-    emergencyContactName: "Clara Rodriguez",
-    emergencyContactPhone: "+1-555-654-3210",
-    emergencyContactRelationship: "Sister",
-    profileTitle: "Home Care Aide",
-    bio: "Experienced home care aide providing compassionate support for seniors.",
-    yearsExperience: "1-2 years",
-    yearsExperienceFloat: 1.5,
-    askingRate: 30.0,
-    rateType: "Hourly",
-    responseTime: "Within 48 hours",
-    servicesOffered: ["Housekeeping", "Companionship"],
-    isProfileComplete: true,
-    lastUpdated: 1708650175742,
-    isPubliclyVisible: true,
-    availability: [
-      "2025-07-20:14",
-      "2025-07-28:09",
-      "2025-07-30:16"
-    ]
-  }, {
-    userId: "user-005",
-    profileOwner: "user-005",
-    firstName: "Ava",
-    firstNameLower: "ava",
-    lastName: "Nguyen",
-    lastNameLower: "nguyen",
-    gender: "Female",
-    languages: ["English", "Vietnamese"],
-    phone: "+1-555-567-8901",
-    email: "ava.nguyen@example.com",
-    preferredContact: "Email",
-    profilePhoto: "/images/home/meet-providers/person-placeholder-10.jpg",
-    address: "234 Cedar Court",
-    city: "Edmonton",
-    province: "Alberta",
-    postalCode: "T5J 2R4",
-    emergencyContactName: "Linh Nguyen",
-    emergencyContactPhone: "+1-555-543-2109",
-    emergencyContactRelationship: "Mother",
-    profileTitle: "Massage Therapist",
-    bio: "Skilled massage therapist specializing in relaxation and pain relief.",
-    yearsExperience: "3-5 years",
-    yearsExperienceFloat: 3.5,
-    askingRate: 55.0,
-    rateType: "Hourly",
-    responseTime: "Within 12 hours",
-    servicesOffered: ["Companionship", "Mobility Assistance"],
-    isProfileComplete: true,
-    lastUpdated: 1708650175742,
-    isPubliclyVisible: true,
-    availability:[
-      "2025-07-26:10",
-      "2025-07-26:11",
-      "2025-07-26:12",
-      "2025-07-27:13",
-      "2025-07-27:14",
-      "2025-07-27:15"
-    ]
-  }
-];
