+++ src/components/marketplace/search-bar.tsx
@@ -15,15 +15,15 @@ const MarketplaceSearchBar = ({searchKey, setSearchKey, triggerFetch}) => {
   return (
     <>
       {/* Navigation Bar */}
-      <div className="flex-col ml-5">
+      <div className="flex-col mg-5">
         <div>
           <h3 className="font-bold text-h4-size md:text-h3-size text-darkest-green mb-2">Find a Caregiver</h3>
         </div>
-        <div className="sm:flex-col sm:items-center lg:flex lg:flex-row lg:justify-between mb-6">
-          <p className="text-body5-size md:text-body3-size">Connect with professional caregivers that match your special needs</p>
-          <div className="relative w-80 ">
-            <Input type="text" value={searchKey} onChange={handleKeyDown} placeholder="Search Caregivers..." className="mt-10 lg:mt-0 pr-10 h-[60px] text-[20px] border-gray-300 rounded-md" onKeyDown={handleKeyDown} />
-            <div className="absolute inset-y-0 right-0 flex items-center pr-3 pointer-events-none">
+        <div className="flex-col items-center lg:flex lg:flex-row lg:justify-between mb-6">
+          <p className="text-body4-size lg:text-body3-size ">Connect with professional caregivers that match your special needs</p>
+          <div className="relative lg:w-80 w-full">
+            <Input type="text" value={searchKey} onChange={handleKeyDown} placeholder="Search Caregivers..." className="mt-2 lg:mt-0 h-[60px] text-[20px] border-gray-300 rounded-md" onKeyDown={handleKeyDown} />
+            <div className="absolute inset-y-0 right-0 flex items-center w-10 justify-center z-10" onClick={triggerFetch}>
               <svg
                 className="h-5 w-5 text-gray-400"
                 xmlns="http://www.w3.org/2000/svg"
